requirements.txt
python-dotenv
groq
opencv-python
pyttsx3
Pillow
gradio
pyserial
Flask